move the Kotlin plugin code to KotlinPluginFacade
