+++ apollo-gradle-plugin/src/main/kotlin/com/apollographql/apollo3/gradle/internal/ApolloPlugin.kt
@@ -1,7 +1,6 @@
 package com.apollographql.apollo3.gradle.internal
 
 import com.apollographql.apollo3.gradle.api.ApolloExtension
-import okio.ByteString.Companion.toByteString
 import org.gradle.api.Plugin
 import org.gradle.api.Project
