+++ apollo-gradle-plugin/src/test/kotlin/com/apollographql/apollo3/gradle/test/JavaOnly.kt
@@ -7,7 +7,7 @@ class JavaOnly {
   @Test
   fun javaOnlyProject() {
     TestUtils.withTestProject("java-only") { dir ->
-      TestUtils.executeTaskAndAssertSuccess( "generateApolloSources", dir)
+      TestUtils.executeTaskAndAssertSuccess(":generateApolloSources", dir)
     }
   }
 }
